% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{absAngle2}
\alias{absAngle2}
\title{Calculate step heading}
\usage{
absAngle2(x)
}
\arguments{
\item{x}{Track generated using the amt package}
}
\value{
numeric value or vector
}
\description{
Function to calculate heading, sometimes also called absolute turning angle
}
