% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calcEndpoints}
\alias{calcEndpoints}
\title{Calculate New Endpoints}
\usage{
calcEndpoints(xy, sl, absta)
}
\arguments{
\item{xy}{Coordinates from where the simulated individual comes from}

\item{sl}{Step length in meters}

\item{absta}{Absolute turning angle}
}
\value{
\code{data.frame}
}
\description{
Function to calculate new endpoints
}
