% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcEndpointsC}
\alias{calcEndpointsC}
\title{Calculate New Endpoints}
\usage{
calcEndpointsC(xy, absta, sl)
}
\arguments{
\item{xy}{\code{matrix} containing xy coordinates}

\item{absta}{absolute turning angle}

\item{sl}{step lengths}
}
\value{
numeric value
}
\description{
Function to calculate new endpoints
}
