% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{closest}
\alias{closest}
\title{Find the Index of A Lagged Closest Date}
\usage{
closest(a, b, lag, unit = "days")
}
\arguments{
\item{a}{vector of dates 1}

\item{b}{vector of dates 2}

\item{lag}{number indicating the lag}

\item{unit}{Unit of the lag}
}
\value{
index of the matched date
}
\description{
Function to identify the index of a lagged closest date
}
