% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{colTrans}
\alias{colTrans}
\title{Create Transparent Colors}
\usage{
colTrans(color, percent = 50)
}
\arguments{
\item{color}{The color that you want to make transparent}

\item{percent}{Value between 0 and 100 indicating how transparent the color
should be}
}
\value{
Hexadecimal color code
}
\description{
Function to create transparent color
}
