% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{createPoints}
\alias{createPoints}
\title{Create Random Points}
\usage{
createPoints(points = NULL, areas = NULL, n = 1, randomize = F)
}
\arguments{
\item{points}{\code{SpatialPointsDataFrame} of the source points}

\item{areas}{\code{SpatialPolygonsDataFrame} of the source areas}

\item{n}{numeric value indicating how many source points should be sampled}

\item{randomize}{logical indicating if source points should be randomized}
}
\value{
\code{SpatialPointsDataFrame}
}
\description{
Function that allows to either select source points in the center of the
catchment area or to randomize source points in the catchment area.
}
