% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{createSegments}
\alias{createSegments}
\title{Create Segments Between Spatial Points}
\usage{
createSegments(x)
}
\arguments{
\item{x}{\code{SpatialPoints} or \code{SpatialPointsDataFrame} containing
points between which spatial lines should be generated}
}
\value{
\code{SpatialLinesDataFrame}
}
\description{
Function to create SpatialLines between \code{SpatialPoints}. In contrast to
\code{spLines} from the \code{raster} package this function returns each line
segment as individual line.
}
