% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{cutLineAfter}
\alias{cutLineAfter}
\title{Cut Lines after Intersection}
\usage{
cutLineAfter(line, polygon, precision = 1)
}
\arguments{
\item{line}{\code{SpatialLines} or \code{SpatialLinesDataFrame} that needs to
be cut}

\item{polygon}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
which is used to cut the line}

\item{precision}{precision at which the lines should be cut. The value needs
to be given in map units and depends on the crs}
}
\value{
\code{SpatialLinesDataFrame}
}
\description{
Function to cut a multiline after the intersection with a desired shape
}
