% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{cutLineClose}
\alias{cutLineClose}
\title{Cut Line at Closest Distance to a Point}
\usage{
cutLineClose(line, point, polygon, precision = 1)
}
\arguments{
\item{line}{\code{SpatialLines} or \code{SpatialLinesDataFrame} that needs to
be cut}

\item{point}{\code{SpatialPoints} or \code{SpatialPointsDataFrame} which is
used as reference to which distances are calculated}

\item{polygon}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
containing polygons within which the lines need to lie before they are cut}

\item{precision}{precision at which the lines should be cut. The value needs
to be given in map units and depends on the crs}
}
\value{
\code{SpatialLinesDataFrame}
}
\description{
Function to cut a spatial line at its closest distance to a spatial point
}
