% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{darken}
\alias{darken}
\title{Darken a color}
\usage{
darken(color, factor = 1.4)
}
\arguments{
\item{color}{A character string naming a color}

\item{factor}{Factor by which the color should be darkened. 1.4 by default}
}
\value{
Hexadecimal code of the darkened color
}
\description{
Function to darken a color
}
\examples{
darken("blue")
plot(1:2, 1:2, cex = 70, pch = 20, col = c("blue", darken("blue", 3)))
}
