% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{degreesToMeters}
\alias{degreesToMeters}
\title{Convert degrees to meters}
\usage{
degreesToMeters(x)
}
\arguments{
\item{x}{A length in degrees}
}
\value{
numeric value or vector
}
\description{
Function to convert degrees to meters (only roughly though)
}
\examples{
degreesToMeters(0.01)
}
