% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{disperse}
\alias{disperse}
\title{Simulate Dispersal}
\usage{
disperse(
  source = NULL,
  covars = NULL,
  model = NULL,
  sl_dist = NULL,
  sl_max = Inf,
  date = as.POSIXct("2015-06-15 03:00:00", tz = "UTC"),
  n_steps = 10,
  n_rsteps = 25,
  stop = F
)
}
\arguments{
\item{source}{\code{SpatialPointsDataFrame} of the source point from which
the disperser should be initiated}

\item{covars}{\code{RasterStack} of the spatial covariates}

\item{model}{iSSF model that was prepared using \code{prepareModel}}

\item{sl_dist}{gamma distribution based on which step lengths should be
sampled}

\item{sl_max}{numeric value indicating the maximal step length (in meters)
that is biologically sensible}

\item{date}{POSIXct of the start date}

\item{n_steps}{number of steps that should be simulated in total}

\item{n_rsteps}{number of random steps proposed at each relocation}

\item{stop}{logical indicating if the simulation should stop when a random
step leaves the study extent. \code{FALSE} by default, indicating that any
random step leaving the study extent will be omitted}
}
\value{

}
\description{
Function to simulate dispersal based on a step selection model that was
fitted in the glmmTMB framework
}
