% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{distanceTo}
\alias{distanceTo}
\title{Calculate Raster Distance}
\usage{
distanceTo(x, value = 1)
}
\arguments{
\item{x}{\code{RasterLayer} on which distances should be calculated}

\item{value}{value to which the distance should be calculated}
}
\value{
\code{RasterLayer}
}
\description{
Function to calculate the distance of a raster cell to the nearest cell
containing a certain value
}
