% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{distanceTraveled}
\alias{distanceTraveled}
\title{Calculate Distance Traveled}
\usage{
distanceTraveled(x, fixes = 6)
}
\arguments{
\item{x}{\code{data.frame} containing steps as prepared in the \code{amt}
package}

\item{fixes}{Numeric value indicaing how many fixes should be considered to
calculate the distance travelled}
}
\value{

}
\description{
Function to identify distance traveled in previous fixes
}
