% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{extendRaster}
\alias{extendRaster}
\title{Extend Raster and Fill with Values}
\usage{
extendRaster(x, y)
}
\arguments{
\item{x}{\code{RasterLayer}}

\item{y}{extent to which the raster should be extended}
}
\value{
\code{RasterLayer}
}
\description{
Function to extend a \code{RasterLayer} an fill the extended area with values
sampled from the original \code{RasterLayer}
}
