% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{extrCov}
\alias{extrCov}
\title{Extract covariates}
\usage{
extrCov(x, y)
}
\arguments{
\item{x}{\code{RasterLayer} or \code{RasterStack} containing covariates}

\item{y}{Spatial feature (\code{sp} package) below which covariates should be
extracted}
}
\value{
\code{data.frame}
}
\description{
Function to extract raster values and calculate their average coverage along
a spatial feature
}
