% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{extrCov3}
\alias{extrCov3}
\title{Extract covariates quickly}
\usage{
extrCov3(raster = NULL, feature = NULL)
}
\arguments{
\item{raster}{\code{velox raster} containing covariates}

\item{feature}{Spatial feature (\code{sp} package) below which covariates
should be extracted}
}
\value{
\code{data.frame}
}
\description{
Function to extract raster values and calculate their average coverage along
a spatial feature. This function requires the velox package which allows a
much quicker value extraction.
}
