% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gCentroidWithin}
\alias{gCentroidWithin}
\title{Create centroid within a polygon}
\usage{
gCentroidWithin(pol)
}
\arguments{
\item{pol}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} in which
the centroids should be identified}
}
\value{
\code{SpatialPointsDataFrame}
}
\description{
Function to create centroid within a polygon
}
