% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getAbsNew}
\alias{getAbsNew}
\title{Calculate New Absolute Turning Angle}
\usage{
getAbsNew(absta = NULL, ta = NULL)
}
\arguments{
\item{absta}{Absolute turning angle}

\item{ta}{Relative turning angle}
}
\value{
numeric value
}
\description{
Function to calculate new absolute turning angle
}
