% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAbsNewC}
\alias{getAbsNewC}
\title{Calculate New Absolute Turning Angle}
\usage{
getAbsNewC(absta, ta)
}
\arguments{
\item{absta}{Absolute turning angle}

\item{ta}{Relative turning angle}
}
\value{
numeric value
}
\description{
Function to calculate new absolute turning angle
}
