% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getCoeffs}
\alias{getCoeffs}
\title{Extract glmmTMB Coefficients}
\usage{
getCoeffs(x, zvalue = FALSE, pvalue = FALSE)
}
\arguments{
\item{x}{glmmTMB model}

\item{zvalue}{Should the value be extracted as well? Needs to be logical.}

\item{pvalue}{Should the pvalue be extracted as well? Needs to be logical.}
}
\value{
\code{data.frame}
}
\description{
Function to Extract Coefficients from a glmmTMB result
}
