% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getFixscheme}
\alias{getFixscheme}
\title{Get Fixrate Scheme of Dog}
\usage{
getFixscheme(data = NULL, dogs = NULL)
}
\arguments{
\item{data}{Data that contains a column entitled "Timestamps"}

\item{dogs}{dogs for which you'd like the fixrate scheme}
}
\value{
\code{data.frame}
}
\description{
Function to obtain the approximate fixrate scheme for a desired dog
}
