% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getMode}
\alias{getMode}
\title{Identify Mode}
\usage{
getMode(v)
}
\arguments{
\item{v}{vector of numeric values for which the mode should be calculated}
}
\value{
numeric value
}
\description{
Function to identify the mode value of a vector
}
