% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{glmm_clogit}
\alias{glmm_clogit}
\title{Fit glmmTMB model}
\usage{
glmm_clogit(formula, data)
}
\arguments{
\item{formula}{formula of the model}

\item{data}{to train the model}
}
\value{
glmmTMB conditional logistic regression model
}
\description{
Write a function to run a glmmTMB conditional logistic regression based on
Muff et al.'s paper
}
