% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interpolatePointsC}
\alias{interpolatePointsC}
\title{Interpolate Between two Points}
\usage{
interpolatePointsC(x1, x2, y1, y2, by)
}
\arguments{
\item{x1}{numeric containing x coordiante of first point}

\item{x2}{numeric containing x coordiante of second point}

\item{y1}{numeric containing y coordiante of first point}

\item{y2}{numeric containing y coordiante of second point}

\item{by}{numeric Approximate distance at which points should be interpolated}
}
\value{
matrix
}
\description{
Function to interpolate coordinates between two points
}
