% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lighten}
\alias{lighten}
\title{Lighten a color}
\usage{
lighten(color, factor = 1.4)
}
\arguments{
\item{color}{A character string naming a color}

\item{factor}{Factor by which the color should be lightened. 1.4 by default}
}
\value{
Hexadecimal code of the lightened color
}
\description{
Function to lighten a color
}
\examples{
lighten("blue")
plot(1:2, 1:2, cex = 70, pch = 20, col = c("blue", lighten("blue", 3)))
}
