% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lineSubstring}
\alias{lineSubstring}
\title{Cut line to given start and endpoint}
\usage{
lineSubstring(line = NULL, start = 0, end = 1)
}
\arguments{
\item{line}{\code{SpatialLines} or \code{SpatialLinesDataFrame} that needs to
be cut}

\item{start}{numeric between 0 and 1. Distance at which the line should start}

\item{end}{numeric between 0 and 1. Distance at which the line should end}
}
\value{
\code{SpatialLines}
}
\description{
Function to cut line to given start and endpoint (as measured in distance
between 0 and 1)
}
