% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lineTrack}
\alias{lineTrack}
\title{Turn steps to Spatial Lines}
\usage{
lineTrack(track, crs, progress = F)
}
\arguments{
\item{track}{Track created using the amt package}

\item{crs}{Coordinate reference system of the track}
}
\value{
\code{SpatialLinesDataFrame}
}
\description{
Function to create spatial lines from steps (from amt package)
}
