% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mccrop}
\alias{mccrop}
\title{Crop Rasterstack on Multiple Cores}
\usage{
mccrop(r, e, ...)
}
\arguments{
\item{r}{\code{RasterStack} to be reclassified}

\item{e}{extent to which the \code{RasterStack} should be cropped}
}
\value{
\code{RasterStack}
}
\description{
Crop a rasterstack quickly using multiple cores
}
