% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mcread_csv}
\alias{mcread_csv}
\title{Read CSVs on Multiple Cores}
\usage{
mcread_csv(x)
}
\arguments{
\item{x}{path of the csvs to be loaded}
}
\value{
\code{tibble}
}
\description{
Function to quickly read a couple of csvs using multiple cores
}
