% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mcreclassify}
\alias{mcreclassify}
\title{Reclassify Rasterstack on Multiple Cores}
\usage{
mcreclassify(r, rcl, ...)
}
\arguments{
\item{r}{\code{RasterStack} to be reclassified}

\item{rcl}{reclassification matrix}
}
\value{
\code{RasterStack}
}
\description{
Function to quickly reclassify rasterstacks on multiple cores
}
