% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{metersToDegrees}
\alias{metersToDegrees}
\title{Convert meters to degrees}
\usage{
metersToDegrees(x)
}
\arguments{
\item{x}{A length in meters}
}
\value{
numeric value or vector
}
\description{
Function to convert meters to degrees (only roughly though)
}
\examples{
metersToDegrees(1)
}
