% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{nearJoin}
\alias{nearJoin}
\title{Join dataframes by date}
\usage{
nearJoin(df1, df2, name1, name2, date1, date2)
}
\arguments{
\item{df1}{"left" data frame}

\item{df2}{"right" data frame}

\item{name1}{Column of df1 containing names of individuals}

\item{name2}{Column of df2 containing names of individuals}

\item{date1}{Column of df1 containing the timestamps}

\item{date2}{Column of df1 containing the timestamps}
}
\value{
\code{data.frame}
}
\description{
Function that allows a left-join by dates
}
