% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{normalizeAngle}
\alias{normalizeAngle}
\title{Normalize Turning Angles}
\usage{
normalizeAngle(x)
}
\arguments{
\item{x}{Turning angle}
}
\value{
numeric value
}
\description{
Function to normalize turning angles to a value between 0 and 2*pi
}
