% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{normalizeMap}
\alias{normalizeMap}
\title{Normalize a Raster}
\usage{
normalizeMap(x)
}
\arguments{
\item{x}{\code{RasterLayer} or \code{RasterStack} that you want to normalize
to a range between 0 and 1}
}
\value{
\code{RasterLayer} or \code{RasterStack}
}
\description{
Function to normalize a map
}
