% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pointsInside}
\alias{pointsInside}
\title{Check if Coordinates are within Extent}
\usage{
pointsInside(xy = NULL, extent = NULL)
}
\arguments{
\item{\code{data.frame}}{or \code{matrix} containing columns representing x
and y coordinates for which you want to know if they lie within an extent or
not.}

\item{\code{list}}{containing xmin, xmax, ymin, and ymax}
}
\value{
Logical (\code{TRUE} = within extent, \code{FALSE} = outside extent)
}
\description{
Function to Check if Points are within or outside an extent
}
