% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{predictScore}
\alias{predictScore}
\title{Predict Step Selection Scores}
\usage{
predictScore(coefficients = NULL, formula = NULL, data = data)
}
\arguments{
\item{coefficients}{named \code{list} containing the coefficients for each
covariate}

\item{formula}{object of class \code{formula}. The same formula that was used
to fit the glmmTMB model}
}
\value{
\code{numeric vector} or \code{numeric value}
}
\description{
Function to predict the step selection scores using a glmmTMB model
}
