% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{prepareCovars}
\alias{prepareCovars}
\title{Prepare Covariates for Simulation}
\usage{
prepareCovars(layers)
}
\arguments{
\item{layers}{\code{RasterStack} of covariates}
}
\value{
\code{list}
}
\description{
Function to prepare covariate layers for dispersal simulation
}
