% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{prepareModel}
\alias{prepareModel}
\title{Prepare Movement Model for Dispersal Simulation}
\usage{
prepareModel(model = NULL)
}
\arguments{
\item{model}{Movement model}
}
\value{
\code{list}
}
\description{
Function to prepare movement model for dispersal simulation
}
