% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rasterizeSpatstat}
\alias{rasterizeSpatstat}
\title{Rasterize Shapes Using spatstat}
\usage{
rasterizeSpatstat(l, r, mc.cores = 1)
}
\arguments{
\item{l}{\code{SpatialPoints}, \code{SpatialLines}, or \code{SpatialPolygons}
to be rasterized}

\item{r}{\code{RasterLayer} onto which the objects should be rasterized}

\item{mc.cores}{numeric How many cores should be used?}
}
\value{
\code{RasterLayer}
}
\description{
Function to rasterize using spatstat
}
