% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rasterizeTerra}
\alias{rasterizeTerra}
\title{Rasterize Shapes Using terra}
\usage{
rasterizeTerra(l, r)
}
\arguments{
\item{l}{\code{SpatialPoints}, \code{SpatialLines}, or \code{SpatialPolygons}
to be rasterized}

\item{r}{\code{RasterLayer} onto which the objects should be rasterized}
}
\value{
\code{RasterLayer}
}
\description{
Function to rasterize using terra
}
