% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rasterizeVelox}
\alias{rasterizeVelox}
\title{Rasterize Shapes Using Velox}
\usage{
rasterizeVelox(l, r)
}
\arguments{
\item{l}{\code{SpatialPoints}, \code{SpatialLines}, or \code{SpatialPolygons}
to be rasterized}

\item{r}{\code{RasterLayer} onto which the objects should be rasterized}
}
\value{
\code{RasterLayer}
}
\description{
Function to rasterize using velox
}
