% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{reprojCoords}
\alias{reprojCoords}
\title{Reproject coordinates to another projection}
\usage{
reprojCoords(xy, from, to)
}
\arguments{
\item{xy}{\code{data.frame} or matrix of coordinates that need to be
transformed.}

\item{from}{CRS of the coordinates}

\item{to}{CRS to which the coordinates should be reprojected}
}
\value{
numeric value or vector
}
\description{
Function to transform coordinates into coordinates of another projection
}
\examples{
degreesToMeters(0.01)
}
