% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{resFix}
\alias{resFix}
\title{Resample GPS fixes}
\usage{
resFix(data, hours, start, tol = 0.5)
}
\arguments{
\item{data}{Data that contains a column entitled "Timestamps"}

\item{hours}{Aspired time between fixes in hours}

\item{start}{Hour at which the resampled tracks should start}

\item{tol}{tolerance (in hours)}
}
\value{
\code{data.frame}
}
\description{
Function to resample fixes to a coarser resolution
}
