% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{resFix2}
\alias{resFix2}
\title{Resample GPS fixes}
\usage{
resFix2(data, hours, start)
}
\arguments{
\item{data}{Data that contains a column entitled "Timestamps"}

\item{hours}{Aspired time between fixes in hours}

\item{start}{Hour at which the resampled tracks should start}
}
\value{
\code{data.frame}
}
\description{
Function to resample fixes to a coarser resolution
}
