% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{scaleCovars}
\alias{scaleCovars}
\title{Scale Covariates}
\usage{
scaleCovars(covars, scaling)
}
\arguments{
\item{covars}{Either a \code{data.frame} or a \code{RasterStack} containing
continuous covariates.}

\item{scaling}{\code{data.frame} containing a "Center" column and a "Scale"
column. The \code{data.frame} must contain rownames to indicating the
covariates to which the scaling parameters belong to.}
}
\value{
\code{data.frame} or \code{RasterStack}
}
\description{
Function to scale covariates using a scaling table
}
