% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{segmentLine}
\alias{segmentLine}
\title{Segment Line into Pieces at Intersections with Polygons}
\usage{
segmentLine(line = NULL, poly = NULL)
}
\arguments{
\item{line}{\code{SpatialLines} or \code{SpatialLinesDataFrame} that needs to
be cut}

\item{poly}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} which
is used to segment the line}
}
\value{
\code{SpatialLines}
}
\description{
Function to segment Line into segments at intersections with polygons
}
