% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{showCoeffs}
\alias{showCoeffs}
\title{Plot Model Results}
\usage{
showCoeffs(
  x,
  shape = 1,
  size = 2,
  whiskers = 1.96,
  xlim = c(-1, 1),
  order = levels(x$Covariate)
)
}
\arguments{
\item{x}{Requires a \code{dataframe} containing the columns Covariate,
Coefficient, and SE}

\item{shape}{shape to be plotted}

\item{size}{size of the shape}

\item{whiskers}{number that indicates the size of the whisikers, i.e. by how
much the SE should be multiplied.}

\item{xlim}{limits on the x-axis}

\item{order}{order in which the coefficients should be plotted}
}
\value{
Plot depicting the model results
}
\description{
Function that allows us to easily depict the coefficients of a model
}
