% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sims2tracks}
\alias{sims2tracks}
\title{Coerce Simulated Trajectories to SpatialLines}
\usage{
sims2tracks(
  simulations = NULL,
  steps = 68,
  sampling = c("Static", "Random"),
  mc.cores = detectCores() - 1
)
}
\arguments{
\item{simulations}{\code{data.frame} resulting from the function
\code{disperse}}

\item{steps}{Number of steps to be considered. Set to 68 by default meaning
that the first 68 steps are considered only}

\item{sampling}{Which tracks should be coerced? One of "Static" or "Random"}

\item{mc.cores}{Number of cores used for the coercion. Set to
\code{detectCores() - 1} by default.}
}
\value{
\code{SpatialLinesDataFrame}
}
\description{
Function to coerce simulated coordinates to lines after a desired number of
steps
}
