% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{splitDat}
\alias{splitDat}
\title{Split data into training and testing sets}
\usage{
splitDat(x, ratio = 0.75)
}
\arguments{
\item{x}{\code{dataframe} to be split}

\item{ratio}{number between 0 and 1 indicating the ratio into which the data
should be split. I.e. if the ratio is set to 0.75, 75\% of the data will end
up as training data}
}
\value{
Named list containing a training dataframes and a testing dataframe
}
\description{
Function to split dataframe into training and testing data
}
