% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{splitLine}
\alias{splitLine}
\title{Split Line into Equal Segments}
\usage{
splitLine(line, regular = T, distance, k = 2)
}
\arguments{
\item{line}{\code{SpatialLines} or \code{SpatialLinesDataFrame} that needs to
be cut}

\item{regular}{a logical value indicating if the line should be cut based on
a regular interval. \code{T} by default. If this is set to \code{F}, the line
is cut into lengths of equal lengths.}
}
\description{
Function to segment a line into equal parts. There are two options. Either
the line is split into equal segments of equal length or the line is split
based on points that are regularly distributed on the line.
}
