% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{splitShape}
\alias{splitShape}
\title{Split Shapefile into List}
\usage{
splitShape(s, n = 2)
}
\arguments{
\item{s}{shape to be split}
}
\value{
List of spatial objects from the \code{sp} package
}
\description{
Function to split an \code{sp} object into a list
}
