% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{splitStack}
\alias{splitStack}
\title{Split RasterStack into List}
\usage{
splitStack(r, n = 2)
}
\arguments{
\item{r}{\code{RasterStack} to be split}

\item{n}{number of splits}
}
\value{
List of \code{RasterStack}
}
\description{
Function to split a \code{RasterStack} into a list
}
