% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{step_met}
\alias{step_met}
\title{Function to calculate step metrics}
\usage{
step_met(x, y, timestamps = NULL, degree = T)
}
\arguments{
\item{x}{x coordinate (should be utm)}

\item{y}{y coordinate (should be utm)}

\item{timestamps}{dates so that the step duration can be computed}

\item{degree}{should angles be converted to degrees?}
}
\value{
dataframe with the step metrics
}
\description{
Function to calculate step metrics (step length, absolute turning angle,
relative turning angle, steptime)
}
