% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{stepBursts}
\alias{stepBursts}
\title{Identify Bursts from Steps}
\usage{
stepBursts(x)
}
\arguments{
\item{x}{Dataframe resulting from the \code{amt} package}
}
\value{
Burst IDs
}
\description{
Function to identify bursts from steps as prepared from the \code{amt}
package
}
