% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{subsims}
\alias{subsims}
\title{Subset simulations randomly}
\usage{
subsims(simulations, nid = NULL, steps = NULL, replace = F)
}
\arguments{
\item{simulations}{\code{data.frame} resulting from the function
\code{disperse}}

\item{nid}{integer Number of randomly drawn trajectories that should be
returned. By default, all trajectories will be returned.}

\item{steps}{integer Number of steps to be considered. By default, all steps
are considered.}
}
\value{
\code{data.frame}
}
\description{
Function to randomly select some of the simulations
}
