% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{visInt2}
\alias{visInt2}
\title{Visualize Interactions}
\usage{
visInt2(model, xVar, yVar, exponentiate = F, norm = T)
}
\arguments{
\item{xVar}{name of the first variable of the interaction}

\item{yVar}{name of the second variable of the interaction}

\item{exponentiate}{logical. Should the predicted values be exponentiated?}

\item{norm}{logical. Should the predicted values be normalized?}

\item{label.color}{color of the axis labels}
}
\value{
\code{RasterLayer}
}
\description{
There is a really nice package that automatically shows how interactions
influence a dependent variable. The package is called "visreg" and the
corresponding function is called visreg2d. However, the package does not work
properly with our model and we need to come up with our own code.
}
