% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{writeForm}
\alias{writeForm}
\title{Prepare model call}
\usage{
writeForm(x, slope = TRUE)
}
\arguments{
\item{x}{covariates that should enter the model}

\item{slope}{logical. Should a random slope be fit to the model?}
}
\value{
formula
}
\description{
Helper function to prepare a glmmTMB model call
}
