% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sims2tracks}
\alias{sims2tracks}
\title{Coerce multiple Simulated Trajectories to SpatialLines}
\usage{
sims2tracks(
  simulations = NULL,
  id = "TrackID",
  crs = NULL,
  keep.data = F,
  mc.cores = 1,
  messages = T
)
}
\arguments{
\item{simulations}{\code{data.frame} resulting from the function
\code{disperse}}

\item{id}{character column name of the column containing the track id}

\item{keep.data}{logical Should the input dataframe be conserved?}

\item{mc.cores}{integer Number of cores used for the coercion. Set to
1 by default.}

\item{messages}{logical Should a progress bar and messages be printed?}
}
\value{
\code{SpatialLinesDataFrame}
}
\description{
Function to coerce simulated coordinates to lines after a desired number of
steps. Basically a wrapper around \code{sim2tracks()}
}
